requirements.txt
comfy==0.0.1
diffusers==0.33.1
numpy==2.3.2
Pillow==11.3.0
torch==2.7.0+cu128
